🎨 - Added unit tests for the `AnimationUtils` class in the `lib/core/animation/animation_utils.

+++ lib/core/animation/animation_utils.dart
@@ -1,10 +1,10 @@
 double lerp(double a, double b, double t) {
-  t = t.clamp(0.0, 1.0) as double;
+  t = t.clamp(0.0, 1.0);
   return a + (b - a) * t;
 }
 
 double easeInOut(double t) {
-  t = t.clamp(0.0, 1.0) as double;
+  t = t.clamp(0.0, 1.0);
   return t * t * (3 - 2 * t);
 }
